requirements.txt
selenium
beautifulsoup4
python-dotenv
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy
